Push has been rudimentarily achieved
Button press and actual application of the force on the enemy has been separated from the on trigger function, which now just detects and attaches the rigidbody of the enemy. IfAttackButtonIsPressedAndEnemyHasBeenDetectedThenPushTheEnemy() is the new function which does the push in the update() and the onTrigger function which previously took care of the detection and application of the force, now just detects and attaches the enemy's rigidbody. This is because onTigger cannot detect input, therefore is only suitable for detection and attaching the detected rigidbody.